AI_Service/requirements.txt
# FastAPI server + dependencies
fastapi==0.115.4
uvicorn[standard]==0.30.6
pydantic==2.9.2
python-dotenv==1.0.1
requests==2.32.3

# AI + generation
google-generativeai==0.7.2
Pillow==11.0.0

# Cloud storage
cloudinary==1.41.0   # replaces AWS S3 / boto3
boto3==1.35.43       # optional; keep if you want fallback S3 support

# Video & media processing
ffmpeg-python==0.2.0

# Optional logging / persistence
pymongo==4.8.0

# For typed hints and helpers
typing-extensions==4.12.2
